Move traitsArray formatting to traits lib file
